+++ client/src/pages/home/Home.jsx
@@ -1,20 +1,23 @@
 import { useState, useEffect } from "react";
 import Navbaruser from '../../components/Navbaruser';
+import FilterStatus from "../Admin/Component/FillterStatus";
 // client/src/components/Navbaruser.jsx
 export const Home = () => {
   const [message, setMessage] = useState("");
 
-  useEffect(() => {
-    fetch("/api")
-      .then((response) => response.json())
-      .then((data) => setMessage(data.message));
-  }, []);
+  // useEffect(() => {
+  //   fetch("/api")
+  //     .then((response) => response.json())
+  //     .then((data) => setMessage(data.message));
+  // }, []);
   // return <div className="mt-10 flex w-full flex-col">{message}</div>;
   return (
     <div>
-      <Navbaruser>
-        
-      </Navbaruser>
+      <Navbaruser></Navbaruser>
+      <div>
+        <FilterStatus></FilterStatus>
+      </div>
     </div>
+ 
   )
 };
